Clean out the Day component
